prettier

+++ contracts/interfaces/IUniswapV2Router.sol
@@ -15,12 +15,12 @@ interface IUniswapV2Router {
   ) external returns (uint256[] memory amounts);
 
   function swapTokensForExactTokens(
-    uint amountOut,
-    uint amountInMax,
+    uint256 amountOut,
+    uint256 amountInMax,
     address[] calldata path,
     address to,
-    uint deadline
-  ) external returns (uint[] memory amounts);
+    uint256 deadline
+  ) external returns (uint256[] memory amounts);
 
   function addLiquidity(
     address tokenA,
